refactor: Simplify Wordle congratulatory message generation
- Removed the redundant `CongratsMessage` component and its associated logic.
- Replaced `getCongrats` utility with a unified `generateCongrats` function in `chat.ts`.
- Integrated `generateCongrats` directly into the Wordle game's mutation success handler to provide real-time congratulatory messages.
- Enhanced `ChatWindow` component:
  - Added `prevMessagesLength` state to optimize scrolling behavior for new messages.
  - Updated `useEffect` to handle scenarios where `messages` are not yet available.
- Cleaned up unused files: `CongratsMessage.tsx` and `getCongrats.ts`.
